M2-6365: fixed PR comments
