+++ src/modules/Builder/consts.ts
@@ -7,13 +7,6 @@ export const DEFAULT_MIN_NUMBER = 1;
 //TODO: remove REACT_HOOK_FORM_KEY_NAME in all places where it was used after release the v.8.0.0 react-hook-form
 export const REACT_HOOK_FORM_KEY_NAME = '_id';
 
-export const observerStyles = {
-  position: 'absolute',
-  height: 'calc(100% + 20rem)',
-  bottom: 0,
-};
-
 export const ITEMS_COUNT_TO_ACTIVATE_STATIC = 20;
 export const REPORTS_COUNT_TO_ACTIVATE_STATIC = 3;
 export const SCORE_CONDS_COUNT_TO_ACTIVATE_STATIC = 2;
-export const SUMMARY_ITEMS_COUNT_TO_ACTIVATE_STATIC = 10;
