+++ src/modules/Dashboard/features/RespondentData/RespondentDataSummary/RespondentDataSummary.const.ts
@@ -1,1 +1,2 @@
 export const DEFAULT_DATE_MAX = new Date(1970, 0, 2, 23, 59, 0);
+export const SUMMARY_ITEMS_COUNT_TO_ACTIVATE_STATIC = 10;
